fix: promise check in range statement (#48)
Range statement initialization requires the number of minimum value
promises to match the number of commitments. There's currently a logic
error that renders this check incorrect.

This PR fixes the logic. A test will be included in a [broader test
PR](https://github.com/tari-project/bulletproofs-plus/pull/47), but the
logic change is here for scope.