fix: reduce vector capacity (#45)
When computing the `d` vector, the prover and verify each allocate too
much space for it. This is unnecessary, since the number of scalar
elements in the vector is known.

This PR fixes the over-allocation by correcting the requested capacity.
This can be checked manually by examining the number of elements pushed
to the vector in both the prover and verifier cases.