test: improve coverage (#47)
[Coverage](https://coveralls.io/github/tari-project/bulletproofs-plus?branch=main)
for the repository is pretty good, but can be improved by adding tests.

This PR adds tests across the board.

There are a few notable changes made for more robust coverage that
should receive extra attention:
- The looping strategy for `Scalar::random_not_zero` has been modified
from a `loop` to a `while`. This imposes an extra constant-time equality
check, but this is done directly on byte arrays and is very fast.
- The `TranscriptProtocol` trait function `append_scalar` was unused, so
it has been removed entirely.
- The `RangeParameter` getter function `bp_gens` was unused, so it has
been removed entirely.