fix: clean up getter functions and remove clones (#63)
The getter functions for the inner-product round vectors have minor
issues.

First, they perform an unnecessary cloning during decompression. Second,
they return a `Result` that produces an error if the vectors are empty,
which cannot happen in practice. Finally, the getters for the compressed
vectors are entirely unnecessary.

This PR removes the cloning, gets rid of the compressed getters, and
changes the return type of the decompressed getters. It also uses
iterators where appropriate.