chore!: clean up `ExtensionDegree` (#85)
Currently, `ExtensionDegree` discriminants are represented as `isize` by
the compiler. The enumeration implements `TryFrom<usize>` via a wrapper
to a public function `ExtensionDegreee::try_from_size` that is used in
the library. This is inelegant, makes deserialization (which acts on
bytes) more awkward, and introduces the possibility that a future
enumeration update could overflow `u8` and cause serialization problems.

This PR moves the functionality of `ExtensionDegree::try_from_size` into
a `TryFrom<u8>` implementation, and implements `TryFrom<usize>` as a
wrapper. It changes the discriminant representation to `u8`, which will
ensure that any future change to `ExtensionDegree` that could overflow
will be caught by the compiler.

BREAKING CHANGE: Introduces a change to the `ExtensionDegree` public
API.