fix: improve prover consistency checks (#98)
When producing a proof, the prover does not currently check that the
provided witness value is within the range specified by the statement
bit length. It also does not check that the witness corresponds to valid
openings of the statement commitments. This is not a security issue (the
verifier will reject such an invalid proof), but it means that the
caller may produce an invalid proof that does not return an error.

This PR adds these checks, and ensures that the prover returns an error
if they fail. It also adds unit tests exercising various failure modes.

Closes #97.