feat: move RNGs into `RangeProofTranscript` (#111)
Recent work in #109 uses Merlin's `TranscriptRng` functionality to
improve the handling of random number generation, and also refactors
transcript operations for safer use through a `RangeProofTranscript`
wrapper.

A
[suggestion](https://github.com/tari-project/bulletproofs-plus/pull/109#discussion_r1449861417)
by @sdbondi recommends moving the `TranscriptRng` into
`RangeProofTranscript` so the prover and verifier aren't responsible for
it. This PR adds such a change.

It also updates `RangeProofTranscript` to hold a mutable reference to
the external random number generator. This ensures the prover and
verifier can't accidentally use it instead of the `TranscriptRng`.