feat!: remove verifier RNG requirement (#116)
The verifier currently requires support for a random number generator,
either provided by the caller or taken from `OsRng`. This is only to
produce weights used in the batch verification process; the weights do
not need to be cryptographically secure, but must not be known to the
prover. 

This PR uses `TranscriptRng` functionality to produce these weights
deterministically. It removes the need for the verifier to supply or
support any random number generator.

This is technically a breaking change, since it removes
`verify_batch_with_rng` from the public API; the existing `verify_batch`
should be used instead, and no longer requires the `rand` feature.

Closes #110.

BREAKING CHANGE: Removes `verify_batch_with_rng` from the public API. No
longer requires that `verify_batch` have the `rand` feature enabled.