fix: reduce verification vector allocation (#127)
Batch verification vectors are allocated in part using the expected
aggregation factor of each statement in the batch. However, this was
being done using the aggregation factors corresponding to each
statement's associated generators, which may exceed the aggregation
factors actually used in the statements. The result was a possible
over-allocation of these vectors.

This PR changes the allocation to use the actual aggregation factors.

Closes #126.