chore: remove use of default for points and scalars (#142)
The use of the `Default` implementation for Ristretto point and scalar
types is somewhat non-idiomatic, since it isn't immediately clear what a
"default" point or scalar must be. In practice, these are the identity
group element and zero scalar, respectively.

This PR removes ambiguity and makes the library more future proof by
removing all such uses. Instead, it uses the `Identity` implementation
for point types and the `ZERO` constant for scalars.

Closes #141.